Cleaning up of Data, N/A, and Treatment Summary
The N/A filter was removed from the reshaping block. The "Group" column was also removed due to it being irrelevant to the summary. Then, once the two sets were joined together, the N/A values were filted out. Finally, summarized the mean, st. dev., and number of values for each consolidated treatment.