Full Table Now appears, small questions, graphs in progress
After implementing more of the code from the referenced work, many entries now load in the table. The temperature values would be aided by clarification and the presence of so many duplicates is troubling, however, The negative value for a temperature entry adds some difficulty to creatively visualizing the data, which is now in progress.